+++ mapmatching/osmnet/combine_edges.py
@@ -1,9 +1,6 @@
-
 import numpy as np
 import pandas as pd
 
-import sys
-sys.path.append('../')
 from ..utils.interval_helper import merge_intervals
 from ..utils.parallel_helper import parallel_process
 
@@ -14,7 +11,6 @@ def calculate_degree(df_edges):
 
     return pd.concat([indegree, outdegree], axis=1).fillna(0).astype(int)
 
-
 def get_aux_points(df_edges, exclude_list=None):
     degree = calculate_degree(df_edges)
     
@@ -23,7 +19,6 @@ def get_aux_points(df_edges, exclude_list=None):
         aux_node_lst = [id for id in aux_node_lst if id not in exclude_list]
 
     return aux_node_lst
-
 
 def combine_links(edges, combine_intervals):
     """Combine OSM links with `rid`.
@@ -60,8 +55,8 @@ def combine_links(edges, combine_intervals):
     
     return edges
 
-
 def pipeline_combine_links(df_edges:pd.DataFrame, exclude_list, n_jobs=8):
+    # BUG multi_edges
     aux_nids = get_aux_points(df_edges, exclude_list=exclude_list)
 
     cands_edges = df_edges.query("src in @aux_nids").sort_values(by=['way_id', 'order'])
@@ -72,7 +67,8 @@ def pipeline_combine_links(df_edges:pd.DataFrame, exclude_list, n_jobs=8):
 
     # parallel process, 不能使用 cands_edges，因为涉及到上下游的合并
     _df_edges = df_edges.query(f"way_id in @cands_way_ids")
-    params = ((df, aux_edge_intervals[i]) for i, df in _df_edges.groupby('way_id'))
+    params = ((df, aux_edge_intervals[i]) 
+                for i, df in _df_edges.groupby('way_id'))
     combined_edges = parallel_process(combine_links, params, pbar_switch=True, 
                                       n_jobs=n_jobs, total=len(cands_way_ids), desc='Combine edges')